third_party/WebKit/LayoutTests/editing/shadow/contenteditable-propagation-to-distributed-node-expected.txt

This test checks that contenteditable is propagated to distributed nodes.

PASS getComputedStyle(shadowRoot.querySelector('span'), null).webkitUserModify is "read-only"
PASS getComputedStyle(shadowRoot.querySelector('content').getDistributedNodes()[0], null).webkitUserModify is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
